docs: update timepicker doc
